+++ bzt/__init__.py
@@ -17,7 +17,7 @@ import os
 import sys
 from abc import abstractmethod
 
-VERSION = "1.7.0"
+VERSION = "1.7.2"
 
 
 class RCProvider(object):