+++ bzt/modules/selenium.py
@@ -29,7 +29,7 @@ from bzt.engine import ScenarioExecutor, Scenario, FileLister, PythonGenerator
 from bzt.modules.aggregator import ConsolidatingAggregator, ResultsReader
 from bzt.modules.console import WidgetProvider, PrioritizedWidget
 from bzt.modules.functional import FunctionalResultsReader, FunctionalAggregator, FunctionalSample
-from bzt.six import string_types, text_type, parse, iteritems
+from bzt.six import string_types, text_type, parse
 from bzt.utils import RequiredTool, shell_exec, shutdown_process, JavaVM, TclLibrary, get_files_recursive
 from bzt.utils import dehumanize_time, MirrorsManager, is_windows, BetterDict, get_full_path
 
@@ -39,7 +39,33 @@ except ImportError:
     from pyvirtualdisplay import Display
 
 
-class SeleniumExecutor(ScenarioExecutor, WidgetProvider, FileLister):
+class AbstractSeleniumExecutor(ScenarioExecutor):
+    """
+    Abstract base class for Selenium executors.
+
+    All executors must implement the following interface.
+    """
+
+    SHARED_VIRTUAL_DISPLAY = {}
+
+    @abstractmethod
+    def add_env(self, env):
+        """
+        Add environment variables into selenium process env
+        :type env: dict[str,str]
+        """
+        pass
+
+    @abstractmethod
+    def get_virtual_display(self):
+        """
+        Return virtual display instance used by this executor.
+        :rtype: Display
+        """
+        pass
+
+
+class SeleniumExecutor(AbstractSeleniumExecutor, WidgetProvider, FileLister):
     """
     Selenium executor
     :type virtual_display: Display
@@ -65,8 +91,6 @@ class SeleniumExecutor(ScenarioExecutor, WidgetProvider, FileLister):
 
     SUPPORTED_TYPES = ["python-nose", "java-junit", "ruby-rspec", "js-mocha"]
 
-    SHARED_VIRTUAL_DISPLAY = {}
-
     def __init__(self):
         super(SeleniumExecutor, self).__init__()
         self.additional_env = {}
@@ -79,6 +103,12 @@ class SeleniumExecutor(ScenarioExecutor, WidgetProvider, FileLister):
         self.self_generated_script = False
         self.generated_methods = BetterDict()
         self.runner_working_dir = None
+
+    def get_virtual_display(self):
+        return self.virtual_display
+
+    def add_env(self, env):
+        self.additional_env.update(env)
 
     def set_virtual_display(self):
         display_conf = self.settings.get("virtual-display")
@@ -378,6 +408,7 @@ class JUnitTester(AbstractTestRunner):
 
         self.base_class_path = [self.selenium_server_jar_path, self.junit_path, self.junit_listener_path,
                                 self.hamcrest_path, self.json_jar_path]
+        self.base_class_path.extend(executor.settings.get("additional-classpath", []))
         self.base_class_path.extend(self.scenario.get("additional-classpath", []))
         self.base_class_path = [os.path.abspath(executor.engine.find_file(x)) for x in self.base_class_path]
 
@@ -1147,14 +1178,23 @@ from selenium.common.exceptions import NoAlertPresentException
             if subject != "body":
                 raise ValueError("Only 'body' subject supported ")
 
+            assert_message = "'%s' " % val
+            if not reverse:
+                assert_message += 'not '
+            assert_message += 'found in BODY'
+
             if regexp:
                 assert_method = "self.assertEqual" if reverse else "self.assertNotEqual"
                 assertion_elements.append(self.gen_statement("re_pattern = re.compile(r'%s')" % val))
-                method = '%s(0, len(re.findall(re_pattern, body)))' % assert_method
+
+                method = '%s(0, len(re.findall(re_pattern, body)), "AssertionError: %s")'
+                method %= assert_method, assert_message
                 assertion_elements.append(self.gen_statement(method))
             else:
                 assert_method = "self.assertNotIn" if reverse else "self.assertIn"
-                assertion_elements.append(self.gen_statement('%s("%s", body)' % (assert_method, val)))
+                method = '%s("%s", body, "AssertionError: %s")'
+                method %= assert_method, val, assert_message
+                assertion_elements.append(self.gen_statement(method))
         return assertion_elements
 
     def __gen_assert_page(self):
@@ -1291,7 +1331,8 @@ class LoadSamplesReader(ResultsReader):
         rcd = self.STATUS_TO_CODE.get(item["status"], "UNKNOWN")
         error = item["error_msg"] if item["status"] in SeleniumReportReader.FAILING_TESTS_STATUSES else None
         trname = ""
-        return tstmp, label, concur, rtm, cnn, ltc, rcd, error, trname
+        byte_count = None
+        return tstmp, label, concur, rtm, cnn, ltc, rcd, error, trname, byte_count
 
     def _read(self, last_pass=False):
         for row in self.report_reader.read(last_pass):