+++ bzt/resources/nose_plugin.py
@@ -94,12 +94,12 @@ class BZTPlugin(Plugin):
         :param error:
         :return:
         """
-        exc_type, exc, trace = error
+        exc_type, exc_msg, trace = error
         # test_dict will be None if startTest wasn't called (i.e. exception in setUp/setUpClass)
         if self.test_dict is not None:
             self.test_dict["status"] = "BROKEN"
-            self.test_dict["error_msg"] = ''.join(traceback.format_exception_only(exc_type, exc)).rstrip()
-            self.test_dict["error_trace"] = ''.join(traceback.format_exception(exc_type, exc, trace)).rstrip()
+            self.test_dict["error_msg"] = exc_msg.split('\n')[0]
+            self.test_dict["error_trace"] = ''.join(traceback.format_exception(exc_type, exc_msg, trace)).rstrip()
 
     def addFailure(self, test, error):  # pylint: disable=invalid-name
         """
@@ -109,10 +109,10 @@ class BZTPlugin(Plugin):
 
         :return:
         """
-        exc_type, exc, trace = error
+        exc_type, exc_msg, trace = error
         self.test_dict["status"] = "FAILED"
-        self.test_dict["error_msg"] = ''.join(traceback.format_exception_only(exc_type, exc)).rstrip()
-        self.test_dict["error_trace"] = ''.join(traceback.format_exception(exc_type, exc, trace)).rstrip()
+        self.test_dict["error_msg"] = exc_msg.split('\n')[0]
+        self.test_dict["error_trace"] = ''.join(traceback.format_exception(exc_type, exc_msg, trace)).rstrip()
 
     def addSkip(self, test):  # pylint: disable=invalid-name
         """
@@ -152,7 +152,7 @@ class BZTPlugin(Plugin):
 def run_nose(report_file, files, iterations, hold):
     argv = [__file__, '-v']
     argv.extend(files)
-    argv.extend(['--with-bzt_plugin', '--nocapture'])
+    argv.extend(['--with-bzt_plugin', '--nocapture', '--exe'])
 
     if iterations == 0:
         if hold > 0: