+++ tests/modules/test_defaultAggregator.py
@@ -20,21 +20,21 @@ class TestDefaultAggregator(BZTestCase):
 
         mock = MockReader()
         mock.buffer_scale_idx = '100.0'
-        mock.data.append((1, "", 1, r(), r(), r(), 200, None, ''))
-        mock.data.append((2, "", 1, r(), r(), r(), 200, None, ''))
-        mock.data.append((2, "", 1, r(), r(), r(), 200, None, ''))
-        mock.data.append((3, "", 1, r(), r(), r(), 200, None, ''))
-        mock.data.append((3, "", 1, r(), r(), r(), 200, None, ''))
-        mock.data.append((4, "", 1, r(), r(), r(), 200, None, ''))
-        mock.data.append((4, "", 1, r(), r(), r(), 200, None, ''))
+        mock.data.append((1, "", 1, r(), r(), r(), 200, None, '', 0))
+        mock.data.append((2, "", 1, r(), r(), r(), 200, None, '', 0))
+        mock.data.append((2, "", 1, r(), r(), r(), 200, None, '', 0))
+        mock.data.append((3, "", 1, r(), r(), r(), 200, None, '', 0))
+        mock.data.append((3, "", 1, r(), r(), r(), 200, None, '', 0))
+        mock.data.append((4, "", 1, r(), r(), r(), 200, None, '', 0))
+        mock.data.append((4, "", 1, r(), r(), r(), 200, None, '', 0))
 
         obj.add_listener(mock)
 
         for point in mock.datapoints():
             self.assertNotEquals(0, point[DataPoint.CUMULATIVE][''][KPISet.CONCURRENCY])
 
-        mock.data.append((2, "", 1, r(), r(), r(), 200, None, ''))
-        mock.data.append((2, "", 1, r(), r(), r(), 200, None, ''))
+        mock.data.append((2, "", 1, r(), r(), r(), 200, None, '', 0))
+        mock.data.append((2, "", 1, r(), r(), r(), 200, None, '', 0))
 
         for point in mock.datapoints():
             pass
@@ -57,7 +57,7 @@ class TestDefaultAggregator(BZTestCase):
         # current measurements shows ~25K samples/sec
         for cnt in (10, 100, 1000, 10000, 25000, 40000, 50000):
             for a in range(0, cnt):
-                sample = (cnt, "", 1, r(1000), r(1000), r(1000), rc(), err(), '')
+                sample = (cnt, "", 1, r(1000), r(1000), r(1000), rc(), err(), '', 0)
                 mock.data.append(sample)
             before = time.time()
             for point in mock.datapoints():
@@ -88,35 +88,35 @@ class TestDefaultAggregator(BZTestCase):
 
         buffer_len = mock.buffer_len
         for i in range(5):
-            mock.data.append((100 + i, "", 1, 2, 2, 2, 200, None, ''))
+            mock.data.append((100 + i, "", 1, 2, 2, 2, 200, None, '', 0))
         points = list(mock.datapoints())
         points = list(mock.datapoints())
         self.assertTrue(mock.buffer_len > buffer_len)
         buffer_len = mock.buffer_len
 
         for i in range(10):
-            mock.data.append((200 + i, "", 1, 3, 3, 3, 200, None, ''))
+            mock.data.append((200 + i, "", 1, 3, 3, 3, 200, None, '', 0))
         points = list(mock.datapoints())
         points = list(mock.datapoints())
         self.assertTrue(mock.buffer_len > buffer_len)
         buffer_len = mock.buffer_len
 
         for i in range(20):
-            mock.data.append((300 + i, "", 1, 4, 4, 4, 200, None, ''))
+            mock.data.append((300 + i, "", 1, 4, 4, 4, 200, None, '', 0))
         points = list(mock.datapoints())
         points = list(mock.datapoints())
         self.assertTrue(mock.buffer_len > buffer_len)
         buffer_len = mock.buffer_len
 
         for i in range(15):
-            mock.data.append((400 + i, "", 1, 1, 1, 1, 200, None, ''))
+            mock.data.append((400 + i, "", 1, 1, 1, 1, 200, None, '', 0))
         points = list(mock.datapoints())
         points = list(mock.datapoints())
         self.assertTrue(mock.buffer_len < buffer_len)
         buffer_len = mock.buffer_len
 
         for i in range(30):
-            mock.data.append((500 + i, "", 1, 1, 1, 1, 200, None, ''))
+            mock.data.append((500 + i, "", 1, 1, 1, 1, 200, None, '', 0))
         points = list(mock.datapoints())
         points = list(mock.datapoints())
         self.assertTrue(mock.buffer_len < buffer_len)