+++ tests/modules/test_locustIOExecutor.py
@@ -5,9 +5,9 @@ import time
 from bzt import six
 from bzt.modules.aggregator import DataPoint, KPISet
 from bzt.modules.locustio import LocustIOExecutor, SlavesReader
+from bzt.modules.provisioning import Local
 from tests import BZTestCase, __dir__
 from tests.mocks import EngineEmul
-from bzt.modules.provisioning import Local
 
 
 class TestLocustIOExecutor(BZTestCase):
@@ -98,6 +98,7 @@ class TestLocustIOExecutor(BZTestCase):
         self.assertEquals(107, len(points))
         for point in points:
             self.assertGreater(point[DataPoint.CURRENT][''][KPISet.AVG_RESP_TIME], 0)
+            self.assertGreater(point[DataPoint.CURRENT][''][KPISet.BYTE_COUNT], 0)
 
     def test_locust_resource_files(self):
         if six.PY3:
@@ -141,7 +142,7 @@ class TestLocustIOExecutor(BZTestCase):
         self.obj.engine.config.merge({
             "execution": [{
                 "executor": "locust",
-                "scenario":{
+                "scenario": {
                     "requests": ["http://blazedemo.com/"]}}]})
 
         self.obj.execution = self.obj.engine.config.get('execution')[0]
@@ -160,6 +161,7 @@ class TestLocustIOExecutor(BZTestCase):
                 "scenario": "loc_sc"}],
             "scenarios": {
                 "loc_sc": {
+                    "keepalive": False,
                     "think-time": "5s",
                     "default-address": "http://blazedemo.com",
                     "headers": {